CRUD Done

+++ controllers/contactController.js
@@ -2,7 +2,7 @@ const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModel");
 //@desc get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler(async (req, res) => {
   const contacts = await Contact.find();
   res.json(contacts);
@@ -10,9 +10,9 @@ const getContacts = asyncHandler(async (req, res) => {
 
 //@desc get contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler(async (req, res) => {
-  const contact = await Contact.findById(req.params.id);
+  const contact = await Contact.findById({ user_id: req.user.id });
   if (!contact) {
     res.status(404);
     throw new Error("Contact Not Found!!");
@@ -22,7 +22,7 @@ const getContact = asyncHandler(async (req, res) => {
 
 //@desc Create New contact
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async (req, res) => {
   console.log("Request Body is :", req.body);
   const { name, email, phone } = req.body;
@@ -35,6 +35,7 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id,
   });
   newContact.save();
   res.status(201).json(newContact);
@@ -42,12 +43,18 @@ const createContact = asyncHandler(async (req, res) => {
 
 //@desc Update the contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 const updateContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("Contact Not Found!!");
+  }
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error(
+      "User dont have permssion to update the other user contacts"
+    );
   }
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
@@ -59,12 +66,18 @@ const updateContact = asyncHandler(async (req, res) => {
 
 //@desc Delete contact
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("Contact Not Found!!");
+  }
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error(
+      "User dont have permssion to delete the other user contacts"
+    );
   }
   await Contact.deleteOne();
   res.status(200).json(contact);
